simplify db model, refactoring
