+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanApplicationDataProviderPort.java
@@ -3,6 +3,5 @@ package com.asc.loanservice.domain.loan.application.port;
 import com.asc.loanservice.infrastructure.repository.model.LoanApplication;
 
 public interface LoanApplicationDataProviderPort {
-
     LoanApplication save(LoanApplication loanApplication);
 }
