--- src/loan-service/src/main/java/com/asc/loanservice/infrastructure/repository/model/Loan.java
@@ -1,23 +1,0 @@
-package com.asc.loanservice.infrastructure.repository.model;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-@Getter
-@Builder
-@Embeddable
-@AllArgsConstructor
-@NoArgsConstructor
-public class Loan {
-
-    private BigDecimal loanAmount;
-    private Integer numberOfInstallments;
-    private LocalDate firstInstallmentDate;
-}
