+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/repository/model/LoanApplication.java
@@ -2,7 +2,6 @@ package com.asc.loanservice.infrastructure.repository.model;
 
 
 
-import com.asc.loanservice.domain.loan.application.LoanEvaluationStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,6 +9,8 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
@@ -22,16 +23,15 @@ public class LoanApplication {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Embedded
-    private Customer customer;
-
-    @Embedded
-    private Loan loan;
-
+    private String customerName;
+    private LocalDate customerBirthday;
+    private String customerTaxId;
+    private BigDecimal customerMonthlyIncome;
+    private BigDecimal loanAmount;
+    private Integer numberOfInstallments;
+    private LocalDate firstInstallmentDate;
     @Enumerated(value = EnumType.STRING)
     private LoanEvaluationStatus loanEvaluationStatus;
-
     @CreationTimestamp
     private LocalDateTime createdDate;
 }
