+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/repository/model/LoanEvaluationStatus.java
@@ -1,0 +1,5 @@
+package com.asc.loanservice.infrastructure.repository.model;
+
+public enum LoanEvaluationStatus {
+    SUCCESS, FAILURE
+}
