+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/rule/port/DebtPort.java
@@ -3,6 +3,5 @@ package com.asc.loanservice.infrastructure.rule.port;
 import com.asc.loanservice.external.DebtCustomerResult;
 
 public interface DebtPort {
-
     DebtCustomerResult checkDebtClient(String customerTaxId);
 }
